Updated improvements from polling without the polling
