+++ Sources/ForageSDK/ForageSDK.swift
@@ -8,49 +8,25 @@
 import VGSCollectSDK
 import Foundation
 
+/**
+ Environment base URL
+ */
 public enum EnvironmentTarget: String {
     case sandbox = "api.sandbox.joinforage.app"
     case cert = "api.cert.joinforage.app"
     case prod = "api.joinforage.app"
 }
 
+/**
+ VGS VaultId
+ */
 private enum VaultId: String {
     case sandbox = "tntagcot4b1"
     case cert = "tntpnht7psv"
     case prod = "tntbcrncmgi"
 }
 
-private enum CardType: String {
-    case ebt = "ebt"
-}
-
-protocol ForageSDKService: AnyObject {
-    var collector: VGSCollect? { get }
-    var service: ForageService? { get }
-    
-    func tokenizeEBTCard(
-        merchantAccount: String,
-        bearerToken: String,
-        completion: @escaping (Result<Data?, Error>) -> Void)
-        
-    func checkBalance(
-        bearerToken: String,
-        merchantAccount: String,
-        paymentMethodReference: String,
-        cardNumberToken: String,
-        completion: @escaping (Result<Data?, Error>) -> Void)
-    
-    func capturePayment(
-        bearerToken: String,
-        merchantAccount: String,
-        paymentReference: String,
-        cardNumberToken: String,
-        completion: @escaping (Result<Data?, Error>) -> Void)
-    
-    func cancelRequest()
-}
-
-public class ForageSDK: ForageSDKService {
+public class ForageSDK {
     
     // MARK: Properties
     
@@ -76,6 +52,12 @@ public class ForageSDK: ForageSDKService {
         self.service = LiveForageService(collector)
     }
     
+    /**
+     ``Config`` struct to set environment(``EnvironmentTarget``) on `ForageSDK` singleton
+     
+    - Parameters:
+       - environment: *EnvironmentTarget* enum to set environment.
+    */
     public struct Config {
         let environment: EnvironmentTarget
 
@@ -84,85 +66,23 @@ public class ForageSDK: ForageSDKService {
         }
     }
     
+    /**
+     Setup ForageSDK using Config struct.
+     
+    - Parameters:
+     - config: *Config* struct object to set environment.
+     
+    ````
+      ForageSDK.setup(
+         ForageSDK.Config(environment: .dev)
+      )
+    ````
+    */
     public class func setup(_ config: Config) {
         ForageSDK.config = config
     }
     
-    // MARK: ForageSDKService Methods
-    
-    public func tokenizeEBTCard(
-        merchantAccount: String,
-        bearerToken: String,
-        completion: @escaping (Result<Data?, Error>) -> Void) {
-        let request = ForagePANRequest(
-            authorization: bearerToken,
-            merchantAccount: merchantAccount,
-            panNumber: panNumber,
-            type: CardType.ebt.rawValue,
-            reusable: true
-        )
-        service?.tokenizeEBTCard(request: request, completion: completion)
-    }
-    
-    public func checkBalance(
-        bearerToken: String,
-        merchantAccount: String,
-        paymentMethodReference: String,
-        cardNumberToken: String,
-        completion: @escaping (Result<Data?, Error>) -> Void) {
-        service?.getXKey(bearerToken: bearerToken) { result in
-            switch result {
-            case .success(let model):
-                let request = ForageBalanceRequest(
-                    authorization: bearerToken,
-                    paymentMethodReference: paymentMethodReference,
-                    cardNumberToken: cardNumberToken,
-                    merchantID: merchantAccount,
-                    xKey: model.alias
-                )
-                self.service?.getBalance(request: request, completion: completion)
-            case .failure(let error):
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    public func capturePayment(
-        bearerToken: String,
-        merchantAccount: String,
-        paymentReference: String,
-        cardNumberToken: String,
-        completion: @escaping (Result<Data?, Error>) -> Void) {
-        service?.getXKey(bearerToken: bearerToken) { result in
-            switch result {
-            case .success(let model):
-                let request = ForageCaptureRequest(
-                    authorization: bearerToken,
-                    paymentReference: paymentReference,
-                    cardNumberToken: cardNumberToken,
-                    merchantID: merchantAccount,
-                    xKey: model.alias
-                )
-                
-                self.service?.requestCapturePayment(request: request, completion: completion)
-            case .failure(let error):
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    public func cancelRequest() {
-        service?.provider.stopRequestOnGoing()
-    }
-    
-    private func getXKey(
-        _ bearerToken: String,
-        completion: @escaping (Result<ForageXKeyModel, Error>) -> Void)
-    -> Void {
-        service?.getXKey(bearerToken: bearerToken) { result in
-            completion(result)
-        }
-    }
+    // MARK: Private Methods
     
     private func vaultID(_ environment: EnvironmentTarget) -> VaultId {
         switch environment {
