--- Sources/ForageSDK/Foundation/Network/Model/ForageBalanceRequest.swift
@@ -1,16 +1,0 @@
-//
-//  ForageBalanceRequest.swift
-//  ForageSDK
-//
-//  Created by Symphony on 23/10/22.
-//
-
-import Foundation
-
-internal struct ForageBalanceRequest {
-    let authorization: String
-    let paymentMethodReference: String
-    let cardNumberToken: String
-    let merchantID: String
-    let xKey: String
-}
