+++ Sources/ForageSDK/Foundation/Network/Model/ForagePANRequestModel.swift
@@ -1,5 +1,5 @@
 //
-//  ForagePANRequest.swift
+//  ForagePANRequestModel.swift
 //  ForageSDK
 //
 //  Created by Symphony on 20/10/22.
@@ -7,7 +7,8 @@
 
 import Foundation
 
-internal struct ForagePANRequest {
+/// `ForagePANRequestModel` used for compose request to tokenize Ebt Card
+internal struct ForagePANRequestModel {
     let authorization: String
     let merchantAccount: String
     let panNumber: String
