+++ Sources/ForageSDK/Foundation/Network/Model/ForageRequestModel.swift
@@ -1,0 +1,17 @@
+//
+//  ForageRetrieveModel.swift
+//  ForageSDK
+//
+//  Created by Symphony on 16/11/22.
+//
+import Foundation
+
+/// `ForageRequestModel` used for compose ForageSDK requests
+internal struct ForageRequestModel {
+    let authorization: String
+    let paymentMethodReference: String
+    let paymentReference: String
+    let cardNumberToken: String
+    let merchantID: String
+    let xKey: String
+}
